Enable nodemon start command
