+++ packages/awtrix/nodemon.json
@@ -1,3 +1,3 @@
 {
-  "watch": ["src/"]
+  "watch": ["./src"]
 }
