+++ packages/awtrix/package.json
@@ -12,7 +12,7 @@
     "build": "npm run build:frontend && npm run build:backend",
     "prepublishOnly": "npm run build",
     "awtrix": "esr ./bin/run.dev.ts",
-    "start": "esr ./bin/run.dev.ts start"
+    "start": "nodemon -e ts --exec 'tsc --noEmit && esr ./bin/run.dev.ts start'"
   },
   "repository": {
     "type": "git",
@@ -43,6 +43,8 @@
     "decompress": "^4.2.1",
     "fs-extra": "^9.0.1",
     "hammerjs": "^2.0.8",
+    "koa": "^2.13.1",
+    "koa-router": "^10.0.0",
     "lodash": "^4.17.20",
     "lowdb": "^1.0.0",
     "node-fetch": "^2.6.0",
@@ -68,14 +70,19 @@
   },
   "devDependencies": {
     "@schemastore/package": "0.0.5",
+    "@types/body-parser": "^1.19.0",
+    "@types/connect": "^3.4.34",
     "@types/decompress": "^4.2.3",
     "@types/fs-extra": "^9.0.6",
     "@types/hammerjs": "^2.0.36",
+    "@types/koa": "^2.13.1",
+    "@types/koa-router": "^7.4.2",
     "@types/lodash": "^4.14.159",
     "@types/lowdb": "^1.0.9",
     "@types/node": "^13.9.1",
     "@types/node-fetch": "^2.5.7",
     "@types/puppeteer": "^3.0.1",
+    "@types/serve-static": "^1.13.9",
     "@types/shortid": "0.0.29",
     "@types/socket.io": "^2.1.11",
     "@types/socket.io-client": "^1.4.34",
