--- packages/awtrix/templates/core/apps/awtrix/0.0.1/backend.js
@@ -1,5 +1,0 @@
-module.exports = (BaseClass) => {
-  return class extends BaseClass {
-    register () { }
-  }
-}
