--- packages/awtrix/templates/core/apps/awtrix/0.0.1/package.json
@@ -1,59 +1,0 @@
-{
-  "name": "awtrix",
-  "version": "0.0.1",
-  "awtrix": {
-    "frontend": true,
-    "backend": true,
-    "assets": true,
-    "settings": {
-      "platform": {
-        "type": "dropdown",
-        "required": true,
-        "options": [
-          "PC",
-          "PSN",
-          "XBOX"
-        ]
-      },
-      "profileName": {
-        "type": "text",
-        "required": true,
-        "pattern": "^[^\\s]+#\\d+$"
-      },
-      "region": {
-        "type": "dropdown",
-        "required": true,
-        "options": [
-          "us",
-          "eu",
-          "asia"
-        ]
-      },
-      "age": {
-        "type": "integer",
-        "min": 0,
-        "max": 150,
-        "default": 25
-      },
-      "heightInMeters": {
-        "type": "decimal",
-        "min": 0,
-        "max": 3,
-        "increment": 0.1,
-        "default": 1.65
-      },
-      "showAvatar": {
-        "type": "boolean",
-        "default": true
-      },
-      "backgroundColor": {
-        "type": "color",
-        "default": "#afafaf",
-        "required": true
-      }
-    }
-  },
-  "dependencies": {
-    "yeelight-node": "^1.1.4"
-  }
-}
