--- packages/awtrix/templates/core/apps/awtrix/0.0.1/translations/de.json
@@ -1,3 +1,0 @@
-{
-  "abc": "Test String"
-}
