--- packages/awtrix/templates/core/apps/awtrix/0.0.1/translations/en.json
@@ -1,8 +1,0 @@
-{
-  "abc": "Test String",
-  "settings.platform.title": "Platform",
-  "settings.platform.description": "The platform you play Overwatch on",
-  "settings.profileName.title": "Profilname",
-  "settings.region.title": "Region",
-  "settings.age.title": "Dein Alter"
-}
