+++ packages/awtrix/templates/core/apps/example/0.1.0/backend.js
@@ -1,0 +1,7 @@
+export default (App) => {
+  return class extends App {
+    register() {
+      console.log(`BackendApp for ${this.config.name} registered.`);
+    }
+  };
+};
