+++ packages/awtrix/templates/core/apps/example/0.1.0/translations/en.json
@@ -1,0 +1,1 @@
+{ }
