+++ packages/example-app/backend.ts
@@ -3,7 +3,7 @@ import { BackendApp } from '@awtrix/common'
 export default (App: typeof BackendApp) => {
   return class extends App {
     register () {
-      console.log("BackendApp for {{name}} registered.")
+      console.log(`BackendApp for ${this.config.name} registered.`)
     }
   }
 }
